Change StateGroupComponent
