+++ Packages/UGF.State/Runtime/StateGroupComponent`1.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0496e50ac87f4147b744b59141fa72a9
+timeCreated: 1711372739
